Rename component to Button
